Update README.md
This GitHub project is known for its state-of-the-art encryption functionality. It is a privacy tool backed by a large community. Primarily, it allows you to send and receive PGP encrypted electronic mails. 

Mailpile’s speedy search engine can handle huge volumes of email data and organize it in a clean web interface. It uses static rules or bayesian classifiers for automatic tagging. Go through the free software and live demos on its website to find out more!